src/apis/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
autopep8==2.0.2
billiard==3.6.4.0
boto3==1.26.127
botocore==1.29.127
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.2.35
cryptography==40.0.2
daphne==4.0.0
Django==4.0.4
django-background-tasks==1.2.5
django-celery-results==2.5.0
django-compat==1.0.15
django-cors-headers==3.14.0
django-storages==1.13.2
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
flower==2.0.1
gunicorn==20.1.0
h11==0.14.0
html2text==2020.1.16
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
MarkupSafe==2.1.2
openapi-codec==1.3.2
packaging==23.1
Pillow==9.5.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2-binary==2.8.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.1.1
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
redis==4.5.4
requests==2.28.2
rq==1.13.0
ruamel.yaml==0.17.21
s3transfer==0.6.1
service-identity==21.1.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.3
# Twisted==22.10.0
# twisted-iocpsupport==1.0.3
# txaio==23.1.1
# typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.4.0
zope.interface==6.0
